Move default vitamin impls into their own package
