+++ bowler-kernel/vitamins/src/main/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitamin/defaultvitamin/DefaultBallBearing.kt
@@ -14,29 +14,25 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.vitamins.vitamin
+package com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin
 
 import com.google.common.collect.ImmutableMap
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.BallBearing
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.CenterOfMass
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.klaxon.KlaxonVitaminTo
 import org.octogonapus.ktguava.klaxon.ConvertImmutableMap
-import org.octogonapus.ktunits.quantities.Angle
-import org.octogonapus.ktunits.quantities.ElectricCurrent
-import org.octogonapus.ktunits.quantities.ElectricPotential
 import org.octogonapus.ktunits.quantities.Length
 import org.octogonapus.ktunits.quantities.Mass
-import org.octogonapus.ktunits.quantities.Torque
 
-data class DefaultStepperMotor(
+data class DefaultBallBearing(
     override val width: Length,
-    override val height: Length,
-    override val boltHoleSpacing: Length,
-    override val bolt: DefaultBolt,
-    override val shaft: DefaultShaft,
-    override val voltage: ElectricPotential,
-    override val holdingTorque: Torque,
-    override val current: ElectricCurrent,
-    override val stepAngle: Angle,
+    override val diameter: Length,
+    override val bore: Length,
     override val mass: Mass,
     override val centerOfMass: CenterOfMass,
     @ConvertImmutableMap
     override val specs: ImmutableMap<String, Any>
-) : StepperMotor
+) : BallBearing, KlaxonVitaminTo {
+
+    override fun toVitamin() = this
+}
