+++ bowler-kernel/vitamins/src/main/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitamin/defaultvitamin/DefaultCapScrew.kt
@@ -14,34 +14,28 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.vitamins.vitamin
+package com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin
 
 import com.google.common.collect.ImmutableMap
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.CapScrew
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.CenterOfMass
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.klaxon.KlaxonVitaminTo
 import org.octogonapus.ktguava.klaxon.ConvertImmutableMap
-import org.octogonapus.ktunits.quantities.AngularVelocity
-import org.octogonapus.ktunits.quantities.ElectricPotential
 import org.octogonapus.ktunits.quantities.Length
 import org.octogonapus.ktunits.quantities.Mass
-import org.octogonapus.ktunits.quantities.Torque
 
-data class DefaultServo(
-    override val width: Length,
-    override val depth: Length,
-    override val height: Length,
-    override val shaftCenterToTopOfBody: Length,
-    override val flangeWidth: Length,
-    override val flangeDepth: Length,
-    override val flangeHeight: Length,
-    override val flangeHeightFromBottomOfBody: Length,
-    override val boltWidthSeparation: Length,
-    override val boltDepthSeparation: Length,
-    override val boltHoleDiameter: Length,
-    override val voltage: ElectricPotential,
-    override val stallTorque: Torque,
-    override val speed: AngularVelocity,
-    override val shaft: Shaft,
+data class DefaultCapScrew(
+    override val keySize: Int,
+    override val keyDepth: Length,
+    override val headDiameter: Length,
+    override val headHeight: Length,
+    override val throughHoleDiameter: Length,
+    override val bodyHeight: Length,
     override val mass: Mass,
     override val centerOfMass: CenterOfMass,
     @ConvertImmutableMap
     override val specs: ImmutableMap<String, Any>
-) : Servo
+) : CapScrew, KlaxonVitaminTo {
+
+    override fun toVitamin() = this
+}
