+++ bowler-kernel/vitamins/src/main/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitamin/defaultvitamin/DefaultTimingBelt.kt
@@ -14,29 +14,26 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.vitamins.vitamin
+package com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin
 
 import com.google.common.collect.ImmutableMap
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.CenterOfMass
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.TimingBelt
 import com.neuronrobotics.bowlerkernel.vitamins.vitamin.klaxon.KlaxonVitaminTo
 import org.octogonapus.ktguava.klaxon.ConvertImmutableMap
-import org.octogonapus.ktunits.quantities.Angle
 import org.octogonapus.ktunits.quantities.Length
 import org.octogonapus.ktunits.quantities.Mass
-import org.octogonapus.ktunits.quantities.Stiffness
 
-data class DefaultTorsionSpring(
-    override val freeLength: Length,
-    override val freeAngle: Angle,
-    override val outerDiameter: Length,
-    override val wireDiameter: Length,
-    override val legLength: Length,
-    override val springConstant: Stiffness,
-    override val maximumLoad: Mass,
+data class DefaultTimingBelt(
+    override val height: Length,
+    override val width: Length,
+    override val toothHeight: Length,
+    override val pitchWidth: Length,
     override val mass: Mass,
     override val centerOfMass: CenterOfMass,
     @ConvertImmutableMap
     override val specs: ImmutableMap<String, Any>
-) : TorsionSpring, KlaxonVitaminTo {
+) : TimingBelt, KlaxonVitaminTo {
 
     override fun toVitamin() = this
 }
