+++ bowler-kernel/vitamins/src/main/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitamin/defaultvitamin/DefaultVexWheel.kt
@@ -14,8 +14,10 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.vitamins.vitamin
+package com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin
 
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.CenterOfMass
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.VexWheel
 import com.neuronrobotics.bowlerkernel.vitamins.vitamin.klaxon.KlaxonVitaminTo
 import org.octogonapus.ktguava.collections.emptyImmutableMap
 import org.octogonapus.ktguava.klaxon.ConvertImmutableMap
@@ -33,7 +35,12 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
          */
         object Omni275 : OmniWheel() {
             override val mass = 0.154.lbM
-            override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+            override val centerOfMass =
+                CenterOfMass(
+                    0.inch,
+                    0.inch,
+                    0.inch
+                )
             @ConvertImmutableMap
             override val specs = emptyImmutableMap<String, Any>()
 
@@ -45,7 +52,12 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
          */
         object Omni325 : OmniWheel() {
             override val mass = 0.132.lbM
-            override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+            override val centerOfMass =
+                CenterOfMass(
+                    0.inch,
+                    0.inch,
+                    0.inch
+                )
             @ConvertImmutableMap
             override val specs = emptyImmutableMap<String, Any>()
 
@@ -57,7 +69,12 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
          */
         object Omni4 : OmniWheel() {
             override val mass = 0.232.lbM
-            override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+            override val centerOfMass =
+                CenterOfMass(
+                    0.inch,
+                    0.inch,
+                    0.inch
+                )
             @ConvertImmutableMap
             override val specs = emptyImmutableMap<String, Any>()
 
@@ -71,7 +88,12 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
          */
         object Wheel275 : TractionWheel() {
             override val mass = 0.11.lbM
-            override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+            override val centerOfMass =
+                CenterOfMass(
+                    0.inch,
+                    0.inch,
+                    0.inch
+                )
             @ConvertImmutableMap
             override val specs = emptyImmutableMap<String, Any>()
 
@@ -83,7 +105,12 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
          */
         object Wheel325 : TractionWheel() {
             override val mass = 0.11.lbM
-            override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+            override val centerOfMass =
+                CenterOfMass(
+                    0.inch,
+                    0.inch,
+                    0.inch
+                )
             @ConvertImmutableMap
             override val specs = emptyImmutableMap<String, Any>()
 
@@ -95,7 +122,12 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
          */
         object Wheel4 : TractionWheel() {
             override val mass = 0.198.lbM
-            override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+            override val centerOfMass =
+                CenterOfMass(
+                    0.inch,
+                    0.inch,
+                    0.inch
+                )
             @ConvertImmutableMap
             override val specs = emptyImmutableMap<String, Any>()
 
@@ -107,7 +139,12 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
          */
         object Wheel5 : TractionWheel() {
             override val mass = 0.352.lbM
-            override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+            override val centerOfMass =
+                CenterOfMass(
+                    0.inch,
+                    0.inch,
+                    0.inch
+                )
             @ConvertImmutableMap
             override val specs = emptyImmutableMap<String, Any>()
 
@@ -120,7 +157,8 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
      */
     object HighTraction : DefaultVexWheel() {
         override val mass = 0.077.lbM
-        override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+        override val centerOfMass =
+            CenterOfMass(0.inch, 0.inch, 0.inch)
         @ConvertImmutableMap
         override val specs = emptyImmutableMap<String, Any>()
 
@@ -132,7 +170,8 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
      */
     object Mecanum : DefaultVexWheel() {
         override val mass = 0.41.lbM
-        override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+        override val centerOfMass =
+            CenterOfMass(0.inch, 0.inch, 0.inch)
         @ConvertImmutableMap
         override val specs = emptyImmutableMap<String, Any>()
 
@@ -144,7 +183,8 @@ sealed class DefaultVexWheel : VexWheel, KlaxonVitaminTo {
      */
     object WheelLeg : DefaultVexWheel() {
         override val mass = 0.074.lbM
-        override val centerOfMass = CenterOfMass(0.inch, 0.inch, 0.inch)
+        override val centerOfMass =
+            CenterOfMass(0.inch, 0.inch, 0.inch)
         @ConvertImmutableMap
         override val specs = emptyImmutableMap<String, Any>()
