+++ bowler-kernel/vitamins/src/test/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitaminsupplier/gitvitaminsupplier/VitaminUtil.kt
@@ -18,20 +18,20 @@ package com.neuronrobotics.bowlerkernel.vitamins.vitaminsupplier.gitvitaminsuppl
 
 import com.google.common.collect.ImmutableMap
 import com.neuronrobotics.bowlerkernel.vitamins.vitamin.CenterOfMass
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultBallBearing
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultBattery
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultBolt
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultCapScrew
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultCompressionSpring
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultDCMotor
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultNut
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultRoundMotor
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultServo
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultShaft
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultStepperMotor
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultTimingBelt
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultTorsionSpring
-import com.neuronrobotics.bowlerkernel.vitamins.vitamin.DefaultVexWheel
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultBallBearing
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultBattery
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultBolt
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultCapScrew
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultCompressionSpring
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultDCMotor
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultNut
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultRoundMotor
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultServo
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultShaft
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultStepperMotor
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultTimingBelt
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultTorsionSpring
+import com.neuronrobotics.bowlerkernel.vitamins.vitamin.defaultvitamin.DefaultVexWheel
 import com.neuronrobotics.bowlerkernel.vitamins.vitamin.VexAngle
 import com.neuronrobotics.bowlerkernel.vitamins.vitamin.VexCChannel
 import com.neuronrobotics.bowlerkernel.vitamins.vitamin.VexMetal
@@ -82,158 +82,169 @@ internal fun <T : Random> T.randomMap(): ImmutableMap<String, Any> =
 internal fun <T : Random> T.randomVexMetal() =
     if (nextBoolean()) VexMetal.ALUMINUM else VexMetal.STEEL
 
-internal fun <T : Random> T.randomBallBearing() = DefaultBallBearing(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomBallBearing() =
+    DefaultBallBearing(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomBattery() = DefaultBattery(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().volt,
-    nextDouble().ampere,
-    nextDouble().coulomb,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomBattery() =
+    DefaultBattery(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().volt,
+        nextDouble().ampere,
+        nextDouble().coulomb,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomBolt() = DefaultBolt(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomBolt() =
+    DefaultBolt(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomCapScrew() = DefaultCapScrew(
-    nextInt(),
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomCapScrew() =
+    DefaultCapScrew(
+        nextInt(),
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomCompressionSpring() = DefaultCompressionSpring(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    Stiffness(nextDouble()),
-    nextDouble().gram,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomCompressionSpring() =
+    DefaultCompressionSpring(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        Stiffness(nextDouble()),
+        nextDouble().gram,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomNut() = DefaultNut(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomNut() =
+    DefaultNut(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomTorsionSpring() = DefaultTorsionSpring(
-    nextDouble().inch,
-    nextDouble().degree,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    Stiffness(nextDouble()),
-    nextDouble().gram,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomTorsionSpring() =
+    DefaultTorsionSpring(
+        nextDouble().inch,
+        nextDouble().degree,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        Stiffness(nextDouble()),
+        nextDouble().gram,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomTimingBelt() = DefaultTimingBelt(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomTimingBelt() =
+    DefaultTimingBelt(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomRoundMotor() = DefaultRoundMotor(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    randomShaft(),
-    randomBolt(),
-    nextDouble().inch,
-    nextDouble().degree,
-    nextDouble().degree,
-    nextDouble().volt,
-    nextDouble().radianPerMinute,
-    nextDouble().ampere,
-    nextDouble().nM,
-    nextDouble().ampere,
-    nextDouble().watt,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomRoundMotor() =
+    DefaultRoundMotor(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        randomShaft(),
+        randomBolt(),
+        nextDouble().inch,
+        nextDouble().degree,
+        nextDouble().degree,
+        nextDouble().volt,
+        nextDouble().radianPerMinute,
+        nextDouble().ampere,
+        nextDouble().nM,
+        nextDouble().ampere,
+        nextDouble().watt,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomDCMotor() = DefaultDCMotor(
-    nextDouble().inch,
-    nextDouble().inch,
-    randomShaft(),
-    nextDouble().inch,
-    nextDouble().inch,
-    randomBolt(),
-    nextDouble().inch,
-    nextDouble().degree,
-    nextDouble().degree,
-    nextDouble().volt,
-    nextDouble().radianPerMinute,
-    nextDouble().ampere,
-    nextDouble().nM,
-    nextDouble().ampere,
-    nextDouble().watt,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomDCMotor() =
+    DefaultDCMotor(
+        nextDouble().inch,
+        nextDouble().inch,
+        randomShaft(),
+        nextDouble().inch,
+        nextDouble().inch,
+        randomBolt(),
+        nextDouble().inch,
+        nextDouble().degree,
+        nextDouble().degree,
+        nextDouble().volt,
+        nextDouble().radianPerMinute,
+        nextDouble().ampere,
+        nextDouble().nM,
+        nextDouble().ampere,
+        nextDouble().watt,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
-internal fun <T : Random> T.randomServo() = DefaultServo(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().volt,
-    nextDouble().nM,
-    nextDouble().radianPerMinute,
-    randomShaft(),
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomServo() =
+    DefaultServo(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().volt,
+        nextDouble().nM,
+        nextDouble().radianPerMinute,
+        randomShaft(),
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
 internal fun <T : Random> T.allShafts() = listOf(
     DefaultShaft.SquareShaft(
@@ -292,20 +303,21 @@ internal fun <T : Random> T.randomShaft() = DefaultShaft.ServoHorn.Arm(
     randomMap()
 )
 
-internal fun <T : Random> T.randomStepperMotor() = DefaultStepperMotor(
-    nextDouble().inch,
-    nextDouble().inch,
-    nextDouble().inch,
-    randomBolt(),
-    randomShaft(),
-    nextDouble().volt,
-    nextDouble().nM,
-    nextDouble().ampere,
-    nextDouble().degree,
-    nextDouble().gram,
-    randomCenterOfMass(),
-    randomMap()
-)
+internal fun <T : Random> T.randomStepperMotor() =
+    DefaultStepperMotor(
+        nextDouble().inch,
+        nextDouble().inch,
+        nextDouble().inch,
+        randomBolt(),
+        randomShaft(),
+        nextDouble().volt,
+        nextDouble().nM,
+        nextDouble().ampere,
+        nextDouble().degree,
+        nextDouble().gram,
+        randomCenterOfMass(),
+        randomMap()
+    )
 
 internal fun <T : Random> T.randomVexCChannel() = VexCChannel(
     randomVexMetal(),
